+++ Age Of TR3WAY - Evolution/Assets/Prefab/BaseAlliée/0%HPA.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e4a9b48aa0a15f4eb53660674663063
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
