+++ Age Of TR3WAY - Evolution/Assets/Prefab/BaseEnnemi/0%HPE.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb500dbf19449a843997a8644d2aa397
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
