+++ Age Of TR3WAY - Evolution/Assets/Script/KillSelf.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class KillSelf : MonoBehaviour {
+
+	// Use this for initialization
+	void Start () {
+        Invoke("DestroySelf", 3);
+    }
+	
+	// Update is called once per frame
+	void DestroySelf () {
+        Destroy(gameObject);
+	}
+}
