+++ Age Of TR3WAY - Evolution/Assets/Script/KillSelf.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41a774d602f4d9248b65c4fa6855bc09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
