+++ Age Of TR3WAY - Evolution/Assets/Script/VictoryAnimationStart.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e43e886745bd75743a24aa34ef3cde29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
