--- Age Of TR3WAY - Evolution/Assets/TextMesh Pro/Examples & Extras.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9dd07ee5a7312674db384cae42b43952
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
