+++ Age Of TR3WAY - Evolution/ProjectSettings/EditorBuildSettings.asset
@@ -6,12 +6,18 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 0
-    path: Assets/Scenes/SampleScene.unity
-    guid: 99c9720ab356a0642a771bea13969a05
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/MainMenu.unity
     guid: b315b184d49ebc344bfc94f9d74ef66d
   - enabled: 1
     path: Assets/Scenes/GameScene.unity
     guid: c69974a43b267f84a978b3e7cc0c0591
+  - enabled: 1
+    path: Assets/Scenes/Scene Victory.unity
+    guid: 469b436e942fadd4db9bb132dbcc01ed
+  - enabled: 1
+    path: Assets/Scenes/Scene Defeat.unity
+    guid: 67d8b6f15f931a748be9c946f4fa4054
   m_configObjects: {}
